Omaha #3477 changed plot info locations to floats

Former-commit-id: 3b97fd15a7a3f5d293859ed086844af660ee9426 [formerly f8a1e79288acbe72e876c449f3ef764397c0743d] [formerly 3b97fd15a7a3f5d293859ed086844af660ee9426 [formerly f8a1e79288acbe72e876c449f3ef764397c0743d] [formerly 8738e9cecec47e090abd6822026187e958b357f7 [formerly 00504d6aa7fefd7b6d28701aae777c0fb7d5ace5]]]
Former-commit-id: 8738e9cecec47e090abd6822026187e958b357f7
Former-commit-id: 90a17fa5a6b0c35df35802357453819680aac2f9 [formerly a180c212c921a6a0bbeb490045c05185d2207ef3]
Former-commit-id: fd5c0ea48158e90826270be92bad5eae3b9ca93e

+++ ncep/gov.noaa.nws.ncep.viz.rsc.plotdata/src/gov/noaa/nws/ncep/viz/rsc/plotdata/rsc/TafPlotInfoRetriever.java
@@ -25,6 +25,7 @@ import com.raytheon.viz.pointdata.rsc.retrieve.AbstractDbPlotInfoRetriever;
  * 11/03/2011              sgurung     Initial creation
  * 11/16/2011              sgurung     Order columns by issue_time and sequenceId desc for correct time matching
  * Jul 23, 2014 3410       bclement    location changed to floats
+ * Aug 08, 2014 3477       bclement    changed plot info locations to floats
  * 
  * </pre>
  * 
@@ -70,8 +71,8 @@ public class TafPlotInfoRetriever extends AbstractDbPlotInfoRetriever {
     protected PlotInfo getPlotInfo(Object[] data) {
         PlotInfo stationInfo = new PlotInfo();
         stationInfo.dataURI = (String) data[0];
-        stationInfo.latitude = ((Number) data[1]).doubleValue();
-        stationInfo.longitude = ((Number) data[2]).doubleValue();
+        stationInfo.latitude = ((Number) data[1]).floatValue();
+        stationInfo.longitude = ((Number) data[2]).floatValue();
         stationInfo.stationId = (String) data[3];
         if (stationInfo.stationId == null) {
             stationInfo.stationId = "" + data[1] + "#" + data[2];
